+++ client/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## 1.12.8
+## [1.12.9](https://github.com/availproject/avail-light/releases/tag/avail-light-client-v1.12.9) - 2025-03-17
+
+- Stop light client if update is available
+- Update `avail-light-core` to 1.2.4
+
+## [1.12.8](https://github.com/availproject/avail-light/releases/tag/avail-light-client-v1.12.8) - 2025-03-06
 
 - Update `avail-light-core` to 1.2.3